+++ src/main/java/work/noice/easya11y/setup/RegisterScheduledScanJobTask.java
@@ -9,31 +9,51 @@ import info.magnolia.jcr.util.PropertyUtil;
 import javax.jcr.Node;
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
+import java.util.Map;
 
 /**
  * Task to register the scheduled accessibility scan job in Magnolia's scheduler configuration.
  */
 public class RegisterScheduledScanJobTask extends AbstractRepositoryTask {
-    
+
+    static final String JOB_PATH = "/modules/scheduler/config/jobs/accessibilityScan";
+
+    static final Map<String, Object> JOB_PROPERTIES = Map.of(
+        "name", "Accessibility Scan",
+        "description", "Automated accessibility scan for all pages",
+        "catalog", "default",
+        "command", "easya11y-serverSideScan",
+        "cron", "0 0 9 ? * MON",
+        "enabled", false
+    );
+
+    static final Map<String, Object> PARAM_PROPERTIES = Map.of(
+        "pagePattern", "/",
+        "wcagLevel", "AA",
+        "maxPages", "50",
+        "sendEmail", "true",
+        "sendDigest", "true"
+    );
+
     public RegisterScheduledScanJobTask() {
-        super("Register scheduled scan job", 
+        super("Register scheduled scan job",
               "Registers the accessibility scan job for scheduled execution");
     }
-    
+
     @Override
     protected void doExecute(InstallContext ctx) throws RepositoryException, TaskExecutionException {
         Session configSession = ctx.getConfigJCRSession();
-        
+
         // Create jobs node if it doesn't exist
         Node schedulerConfig = configSession.getNode("/modules/scheduler/config");
         Node jobsNode;
-        
+
         if (!schedulerConfig.hasNode("jobs")) {
             jobsNode = schedulerConfig.addNode("jobs", NodeTypes.ContentNode.NAME);
         } else {
             jobsNode = schedulerConfig.getNode("jobs");
         }
-        
+
         // Create the scheduled scan job configuration
         Node jobNode;
         if (jobsNode.hasNode("accessibilityScan")) {
@@ -41,17 +61,12 @@ public class RegisterScheduledScanJobTask extends AbstractRepositoryTask {
         } else {
             jobNode = jobsNode.addNode("accessibilityScan", NodeTypes.ContentNode.NAME);
         }
-        
+
         // Set job properties
-        PropertyUtil.setProperty(jobNode, "active", false); // Disabled by default
-        PropertyUtil.setProperty(jobNode, "name", "Accessibility Scan");
-        PropertyUtil.setProperty(jobNode, "description", "Automated accessibility scan for all pages");
-        PropertyUtil.setProperty(jobNode, "catalogName", "default");
-        PropertyUtil.setProperty(jobNode, "jobName", "accessibilityScan");
-        PropertyUtil.setProperty(jobNode, "command", "easya11y-serverSideScan");
-        PropertyUtil.setProperty(jobNode, "cron", "0 0 9 ? * MON"); // Every Monday at 9 AM
-        PropertyUtil.setProperty(jobNode, "enabled", false);
-        
+        for (Map.Entry<String, Object> property : JOB_PROPERTIES.entrySet()) {
+            PropertyUtil.setProperty(jobNode, property.getKey(), property.getValue());
+        }
+
         // Create params node for job parameters
         Node paramsNode;
         if (jobNode.hasNode("params")) {
@@ -59,14 +74,12 @@ public class RegisterScheduledScanJobTask extends AbstractRepositoryTask {
         } else {
             paramsNode = jobNode.addNode("params", NodeTypes.ContentNode.NAME);
         }
-        
+
         // Set default job parameters
-        PropertyUtil.setProperty(paramsNode, "pagePattern", "/");
-        PropertyUtil.setProperty(paramsNode, "wcagLevel", "AA");
-        PropertyUtil.setProperty(paramsNode, "maxPages", "50");
-        PropertyUtil.setProperty(paramsNode, "sendEmail", "true");
-        PropertyUtil.setProperty(paramsNode, "sendDigest", "true");
-        
+        for (Map.Entry<String, Object> property : PARAM_PROPERTIES.entrySet()) {
+            PropertyUtil.setProperty(paramsNode, property.getKey(), property.getValue());
+        }
+
         ctx.info("Registered scheduled accessibility scan job");
     }
 }